# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/deviceFeeds": {
            "get": {
                "tags": [
                    "DeviceFeeds"
                ],
                "summary": "Get a listing of the DeviceFeeds.",
                "description": "Get all DeviceFeeds",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/DeviceFeeds"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "DeviceFeeds"
                ],
                "summary": "Store a newly created DeviceFeeds in storage",
                "description": "Store DeviceFeeds",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "DeviceFeeds that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/DeviceFeeds"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/DeviceFeeds"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/deviceFeeds/{id}": {
            "get": {
                "tags": [
                    "DeviceFeeds"
                ],
                "summary": "Display the specified DeviceFeeds",
                "description": "Get DeviceFeeds",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of DeviceFeeds",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/DeviceFeeds"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "DeviceFeeds"
                ],
                "summary": "Update the specified DeviceFeeds in storage",
                "description": "Update DeviceFeeds",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of DeviceFeeds",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "DeviceFeeds that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/DeviceFeeds"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/DeviceFeeds"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "DeviceFeeds"
                ],
                "summary": "Remove the specified DeviceFeeds from storage",
                "description": "Delete DeviceFeeds",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of DeviceFeeds",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "DeviceFeeds": {
            "required": [
                ""
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "humidity": {
                    "description": "humidity",
                    "type": "number",
                    "format": "float"
                },
                "temperature": {
                    "description": "temperature",
                    "type": "number",
                    "format": "float"
                },
                "device_id": {
                    "description": "device_id",
                    "type": "integer",
                    "format": "int32"
                },
                "serial_number": {
                    "description": "device_serial_number",
                    "type": "string"
                }
            }
        },
...
    }
}